Report file - first iteration
This is the first iteration of our report file. It started as a copy of our proposal (since it'll likely share the same content in the beginning), and then I added a section "Creating the KNN Classifier" where I created a GridSearch to find the optimal K value for our model.